+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 1.2.0 (2019-08-20)
+
+
+### Features
+
+* 完成 1.0.0 版本 ([5d3dd09](https://github.com/liuweiGL/amap-echarts/commit/5d3dd09))
+* **amap-charts:** 调整插件设计，允许每个实例进行单独配置 ([543b250](https://github.com/liuweiGL/amap-echarts/commit/543b250))
+* **amap-echarts:** 事件修改、完善插件 ([a497f6f](https://github.com/liuweiGL/amap-echarts/commit/a497f6f))
+* **amap-echarts:** 调整 api 命名、导出接口 ([c171e8a](https://github.com/liuweiGL/amap-echarts/commit/c171e8a))
+* **event:** 调整事件触发时机，分离 render、update 事件 ([99e06b7](https://github.com/liuweiGL/amap-echarts/commit/99e06b7))
+* **plugins:** 添加 AutoVisible 插件 ([a23ebca](https://github.com/liuweiGL/amap-echarts/commit/a23ebca))
+* **开发:** 修改 demo ([f0c5f8d](https://github.com/liuweiGL/amap-echarts/commit/f0c5f8d))
+
 ## 1.1.0 (2019-08-20)